Add NoOp client provider to allow no-operation scaling evaluation.
When initially implementing Chemtrail within an environment, it is
normal for users to be wary of the impacts and changes that will be
made. The NoOp provider solves this by allowing Chemtrail to
perform all its checks and calculations without altering the state
of the Nomad cluster. Chemtrail will even go through its event
cycle which will provide operators with great information.