Fix panic when node is running an old version of the Nomad client.
Since v0.9.0 the Nomad node API uses the NodeResources and
ReservedResources fields to track the nodes available resource for
running work. Chemtrail was blindly assuming these existed which
meant it would panic when attempting to handle a node running a
version of Nomad pre-0.9.0. This change adds compatibility and
checks that ensure Chemtrail works with versions either side of
this change without panics.